Sticky footer and country list table header
To improve table readability particularly for first time consumers of the country plant list page the table header will stick to the top when scrolling down. This only affects non-mobile views since on mobile a non table layout is used.
Also added on desktop screens the footer to be sticky to improve visibilty of export options and OSM license